# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Blood & Emergency Donor Network API
  version: 1.0.0
  description: Recommended hackathon API contract for the smart blood donor network.
servers:
  - url: http://localhost:8000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: integer}
...
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, phone, password, role]
      properties:
        name: {type: string, example: Ali Ahmed}
        email: {type: string, format: email, example: ali@example.com}
        phone: {type: string, example: "+923001234567"}
        password: {type: string, format: password}
        role:
          type: string
          enum: [ADMIN, REQUESTER, DONOR, HOSPITAL]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    DonorRequest:
      type: object
      required: [user_id, blood_group, date_of_birth, city, latitude, longitude]
      properties:
        user_id: {type: integer}
        blood_group: {type: string, example: B+}
        date_of_birth: {type: string, format: date}
        city: {type: string, example: Jamshoro}
        latitude: {type: number, format: double}
        longitude: {type: number, format: double}
        is_available: {type: boolean, default: true}
        is_eligible: {type: boolean, default: true}
        last_donation_date: {type: string, format: date, nullable: true}
    HospitalRequest:
      type: object
      required: [name, address, city, latitude, longitude]
      properties:
        name: {type: string}
        address: {type: string}
        city: {type: string}
        phone: {type: string}
        latitude: {type: number, format: double}
        longitude: {type: number, format: double}
...
    AnalyzeRequest:
      type: object
      required: [request_id, description]
      properties:
        request_id: {type: integer}
        description: {type: string}
...
    DonationCreate:
      type: object
      required: [request_id, donor_id, units]
      properties:
        request_id: {type: integer}
        donor_id: {type: integer}
        units: {type: integer, minimum: 1}
    ConfirmDonation:
      type: object
      required: [confirmation_method]
      properties:
        confirmation_method:
          type: string
          enum: [MANUAL, HOSPITAL, QR]
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201': {description: Registered successfully}
        '400': {description: Invalid request}
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: Login successful}
        '401': {description: Invalid credentials}
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current user}
  /api/donors:
    get:
      tags: [Donors]
      summary: List donors
      security: [{bearerAuth: []}]
      parameters:
        - {name: blood_group, in: query, schema: {type: string}}
        - {name: city, in: query, schema: {type: string}}
        - {name: available, in: query, schema: {type: boolean}}
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Donor list}
    post:
      tags: [Donors]
      summary: Create donor profile
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DonorRequest'}
      responses:
        '201': {description: Donor created}
...
  /api/hospitals:
    get:
      tags: [Hospitals]
      summary: List hospitals
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Hospital list}
    post:
      tags: [Hospitals]
      summary: Create hospital
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HospitalRequest'}
      responses:
        '201': {description: Hospital created}
  /api/hospitals/{id}:
    get:
      tags: [Hospitals]
      summary: Get hospital
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {description: Hospital details}
...
  /api/ai/analyze-request:
    post:
      tags: [AI]
      summary: Analyze urgency and summarize request
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AnalyzeRequest'}
      responses:
        '200': {description: AI analysis}
  /api/ai/check-duplicate:
    post:
      tags: [AI]
      summary: Check duplicate request
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request_id]
              properties:
                request_id: {type: integer}
      responses:
        '200': {description: Duplicate analysis}
...
  /api/donations:
    post:
      tags: [Donations]
      summary: Create donation
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DonationCreate'}
      responses:
        '201': {description: Donation scheduled}
  /api/donations/{id}/confirm:
    patch:
      tags: [Donations]
      summary: Confirm completed donation
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConfirmDonation'}
      responses:
        '200': {description: Donation confirmed}
  /api/donations/{id}/qr:
    post:
      tags: [Donations]
      summary: Generate QR confirmation token
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {description: QR token generated}
  /api/donations/confirm-qr:
    post:
      tags: [Donations]
      summary: Confirm donation using QR token
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qr_token]
              properties:
                qr_token: {type: string}
      responses:
        '200': {description: QR donation confirmed}
  /api/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Dashboard summary
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Summary metrics}
  /api/dashboard/blood-groups:
    get:
      tags: [Dashboard]
      summary: Blood group statistics
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Blood group statistics}
...